requirements.txt
playwright
beautifulsoup4
pandas
openpyxl
boto3
python-dotenv
